move all platform dependent code to their respective dirs
- cleanup platform.h
- no more need for gl_fwd.h, moved to gl.h
- only PLATFORM_ANDROID needs some special handling in gl.h
