+++ core/src/gl/error.cpp
@@ -1,5 +1,4 @@
 #include "error.h"
-#include "gl.h"
 
 namespace Tangram {
