+++ core/src/gl/error.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "platform.h"
-#include "gl/gl_fwd.h"
+#include "gl.h"
 #include <string>
 #include <unordered_map>
