+++ core/src/gl/primitives.cpp
@@ -8,7 +8,6 @@
 #include "gl/vertexLayout.h"
 #include "gl/renderState.h"
 #include "platform.h"
-#include "gl.h"
 
 namespace Tangram {
