+++ core/src/gl/renderState.cpp
@@ -1,6 +1,5 @@
 #include "renderState.h"
 
-#include "gl.h"
 #include "platform.h"
 
 namespace Tangram {
