+++ core/src/gl/shaderProgram.cpp
@@ -1,6 +1,5 @@
 #include "shaderProgram.h"
 
-#include "gl.h"
 #include "platform.h"
 #include "scene/light.h"
 #include "gl/renderState.h"
