+++ core/src/gl/texture.cpp
@@ -1,6 +1,5 @@
 #include "texture.h"
 
-#include "gl.h"
 #include "platform.h"
 #include "util/geom.h"
 #include "gl/renderState.h"
