+++ core/src/gl/textureCube.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "texture.h"
-#include "gl/gl_fwd.h"
+#include "gl.h"
 
 #include <vector>
 #include <string>
