+++ core/src/gl/vboMesh.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gl/gl_fwd.h"
+#include "gl.h"
 #include "vertexLayout.h"
 #include "util/types.h"
