+++ core/src/gl/vertexLayout.cpp
@@ -1,7 +1,6 @@
 #include "vertexLayout.h"
 #include "shaderProgram.h"
 #include "platform.h"
-#include "gl.h"
 
 namespace Tangram {
