+++ core/src/gl/vertexLayout.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gl/gl_fwd.h"
+#include "gl.h"
 
 #include <vector>
 #include <unordered_map>
