--- core/src/platform_gl.h
@@ -1,41 +1,0 @@
-#pragma once
-#ifdef UNIT_TESTS
-#include "gl/gl_fwd.h"
-#else
-
-#ifdef PLATFORM_ANDROID
-#include <GLES2/gl2.h>
-#include <GLES2/gl2ext.h>
-#endif
-
-#ifdef PLATFORM_IOS
-#include <OpenGLES/ES2/gl.h>
-#include <OpenGLES/ES2/glext.h>
-#endif
-
-#ifdef PLATFORM_OSX
-#define GLFW_INCLUDE_GLEXT
-#include <GLFW/glfw3.h>
-/*
- * typedef to resolve name conflict in osx
- */
-#define glClearDepthf glClearDepth
-#define glDepthRangef glDepthRange
-#endif
-
-#ifdef PLATFORM_LINUX
-#define GL_GLEXT_PROTOTYPES
-#include <GLFW/glfw3.h>
-#endif
-
-#ifdef PLATFORM_RPI
-//  Broadcom hardware library for hijacking the GPU card without window manager
-//
-#include "bcm_host.h"
-
-#include <GLES2/gl2.h>
-#include <EGL/egl.h>
-#include <EGL/eglext.h>
-#endif
-
-#endif // !UNIT_TESTS
