+++ core/src/text/textBuffer.cpp
@@ -1,6 +1,5 @@
 #include "textBuffer.h"
 
-#include "gl.h"
 #include "labels/textLabel.h"
 #include "gl/texture.h"
 #include "gl/vboMesh.h"
