+++ ios/src/platform_ios.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "platform.h"
+
+#include <OpenGLES/ES2/gl.h>
+#include <OpenGLES/ES2/glext.h>
+
+#ifdef __OBJC__
+#import "ViewController.h"
+void init(ViewController* _controller);
+#endif
