+++ linux/src/platform_linux.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "platform.h"
+
+#define GL_GLEXT_PROTOTYPES
+#include <GLFW/glfw3.h>
+
+void processNetworkQueue();
+
