+++ osx/src/platform_osx.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "platform.h"
+
+#define GLFW_INCLUDE_GLEXT
+#include <GLFW/glfw3.h>
+
+void NSurlInit();
