+++ rpi/src/platform_rpi.h
@@ -1,0 +1,11 @@
+#pragma once
+
+//  Broadcom hardware library for hijacking the GPU card without window manager
+//
+#include "bcm_host.h"
+
+#include <GLES2/gl2.h>
+#include <EGL/egl.h>
+#include <EGL/eglext.h>
+
+void processNetworkQueue();
