+++ toolchains/android.cmake
@@ -9,31 +9,36 @@ else()
 endif()
 
 # configurations
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -std=c++1y -pedantic -llog -lz -landroid")
-set(CXX_FLAGS_DEBUG "${CXX_FLAGS_DEBUG} -g -O0")
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -fPIC")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++1y -pedantic")
 
-# For CMake 3.0
 if (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
     # using regular Clang or AppleClang
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-gnu-anonymous-struct -Wno-nested-anon-types -Wno-gnu-zero-variadic-macro-arguments")
 endif()
 
-# adding manually the two jni c++ files
-set(ADDITIONAL_TARGET_DEPENDENT_SRC_FILES
-     ${CMAKE_SOURCE_DIR}/android/tangram/jni/jniExports.cpp
-     ${CMAKE_SOURCE_DIR}/android/tangram/jni/jniGenerated.cpp
-     ${CMAKE_SOURCE_DIR}/android/tangram/jni/platform_android.cpp)
-
 # load core library
-set(INSTALL_CORE_LIBRARY "ON")
-set(CORE_LIB_TYPE SHARED)
-set(CORE_INSTALLATION_PATH ${CMAKE_SOURCE_DIR}/android/tangram/libs/${ANDROID_ABI})
-set(CORE_LIB_DEPS GLESv2)
-set(CORE_LIB_NAME tangram) # in order to have libtangram.so
-
 add_subdirectory(${PROJECT_SOURCE_DIR}/core)
-include_directories(${CORE_INCLUDE_DIRS})
+
+set(LIB_NAME tangram) # in order to have libtangram.so
+
+add_library(${LIB_NAME} SHARED
+  ${CMAKE_SOURCE_DIR}/android/tangram/jni/jniExports.cpp
+  ${CMAKE_SOURCE_DIR}/android/tangram/jni/jniGenerated.cpp
+  ${CMAKE_SOURCE_DIR}/android/tangram/jni/platform_android.cpp)
+
+target_link_libraries(${LIB_NAME}
+  PUBLIC
+  core
+  # android libaries
+  GLESv2 log z android)
+
+target_compile_options(${LIB_NAME}
+  PUBLIC
+   -fPIC)
+
+# install to android library dir
+set(LIB_INSTALLATION_PATH ${CMAKE_SOURCE_DIR}/android/tangram/libs/${ANDROID_ABI})
+install(TARGETS ${LIB_NAME} DESTINATION ${LIB_INSTALLATION_PATH})
 
 # link and build functions
 function(link_libraries)
